Fix undo for paste
