+++ app/electron/main/code/diff/helpers.ts
@@ -1,3 +1,5 @@
+import * as t from '@babel/types';
+import { EditorAttributes } from '/common/constants';
 import { CodeDiffRequest } from '/common/models/code';
 import { TemplateNode } from '/common/models/element/templateNode';
 
@@ -14,4 +16,18 @@ export function createHashedTemplateToCodeDiff(
 
 export function hashTemplateNode(node: TemplateNode): string {
     return `${node.path}:${node.startTag.start.line}:${node.startTag.start.column}:${node.startTag.end.line}:${node.startTag.end.column}`;
+}
+
+export function addKeyToElement(element: t.JSXElement): void {
+    if (t.isJSXElement(element)) {
+        const keyExists =
+            element.openingElement.attributes.findIndex(
+                (attr) => t.isJSXAttribute(attr) && attr.name.name === 'key',
+            ) !== -1;
+        if (!keyExists) {
+            const key = EditorAttributes.ONLOOK_MOVE_KEY_PREFIX + Date.now().toString();
+            const keyAttribute = t.jsxAttribute(t.jsxIdentifier('key'), t.stringLiteral(key));
+            element.openingElement.attributes.push(keyAttribute);
+        }
+    }
 }
