+++ app/electron/main/code/diff/move.ts
@@ -1,8 +1,7 @@
 import { NodePath } from '@babel/traverse';
 import * as t from '@babel/types';
 import { getTemplateNode } from '../templateNode';
-import { hashTemplateNode } from './helpers';
-import { EditorAttributes } from '/common/constants';
+import { addKeyToElement, hashTemplateNode } from './helpers';
 import { CodeMove } from '/common/models/actions/code';
 
 export function moveElementInNode(
@@ -24,7 +23,7 @@ export function moveElementInNode(
 
     if (elementToMoveIndex !== -1) {
         const [elementToMove] = children.splice(elementToMoveIndex, 1);
-        addMoveKeyToElement(elementToMove as t.JSXElement);
+        addKeyToElement(elementToMove as t.JSXElement);
         const jsxElements = children.filter(
             (child) => t.isJSXElement(child) || t.isJSXFragment(child) || child === elementToMove,
         ) as t.JSXElement[];
@@ -40,19 +39,5 @@ export function moveElementInNode(
         }
     } else {
         console.error('Element to be moved not found');
-    }
-}
-
-function addMoveKeyToElement(element: t.JSXElement): void {
-    if (t.isJSXElement(element)) {
-        const keyExists =
-            element.openingElement.attributes.findIndex(
-                (attr) => t.isJSXAttribute(attr) && attr.name.name === 'key',
-            ) !== -1;
-        if (!keyExists) {
-            const key = EditorAttributes.ONLOOK_MOVE_KEY_PREFIX + Date.now().toString();
-            const keyAttribute = t.jsxAttribute(t.jsxIdentifier('key'), t.stringLiteral(key));
-            element.openingElement.attributes.push(keyAttribute);
-        }
     }
 }
