+++ app/electron/preload/webview/elements/dom/insert.ts
@@ -79,6 +79,7 @@ export function insertElement(
 function createElement(element: ActionElement) {
     const newEl = document.createElement(element.tagName);
     newEl.setAttribute(EditorAttributes.DATA_ONLOOK_INSERTED, 'true');
+    newEl.removeAttribute(EditorAttributes.DATA_ONLOOK_ID);
 
     for (const [key, value] of Object.entries(element.attributes)) {
         newEl.setAttribute(key, value);
